+++ src/main/java/dabang/star/cafe/domain/admin/OfficeAdminServiceImpl.java
@@ -36,12 +36,12 @@ public class OfficeAdminServiceImpl implements OfficeAdminService {
     }
 
     @Override
-    public void deleteOffice(Long officeId) {
+    public void deleteOffice(Integer officeId) {
         officeRepository.deleteById(officeId);
     }
 
     @Override
-    public void updateOffice(Long id, String name, String address, BigDecimal latitude, BigDecimal longitude) {
+    public void updateOffice(Integer id, String name, String address, BigDecimal latitude, BigDecimal longitude) {
 
         Office office = officeRepository.findById(id).orElseThrow(
                 () -> new OfficeNotFoundException("office not found")
